WE-648 Refactor navigationStateReducer
