+++ Src/WitsmlExplorer.Frontend/components/ContentView.tsx
@@ -1,7 +1,6 @@
 import React, { useContext, useEffect, useState } from "react";
 import styled from "styled-components";
-import NavigationContext from "../contexts/navigationContext";
-import { selectedJobsFlag } from "../contexts/navigationStateReducer";
+import NavigationContext, { selectedJobsFlag } from "../contexts/navigationContext";
 import { BhaRunsListView } from "./ContentViews/BhaRunsListView";
 import { CurveValuesView } from "./ContentViews/CurveValuesView";
 import JobsView from "./ContentViews/JobsView";
