+++ Src/WitsmlExplorer.Frontend/components/ContextMenus/BhaRunContextMenu.tsx
@@ -1,7 +1,7 @@
 import { Typography } from "@equinor/eds-core-react";
 import { Divider, MenuItem } from "@material-ui/core";
 import React from "react";
-import { UpdateWellboreBhaRunsAction } from "../../contexts/navigationStateReducer";
+import { UpdateWellboreBhaRunsAction } from "../../contexts/modificationActions";
 import { DisplayModalAction, HideContextMenuAction, HideModalAction } from "../../contexts/operationStateReducer";
 import OperationType from "../../contexts/operationType";
 import { ObjectType } from "../../models/objectType";
