+++ Src/WitsmlExplorer.Frontend/components/ContextMenus/LogObjectContextMenu.tsx
@@ -2,8 +2,8 @@ import { Typography } from "@equinor/eds-core-react";
 import { Divider, ListItemIcon, makeStyles, MenuItem } from "@material-ui/core";
 import { ImportExport } from "@material-ui/icons";
 import React from "react";
+import { UpdateWellboreLogAction, UpdateWellboreLogsAction } from "../../contexts/modificationActions";
 import ModificationType from "../../contexts/modificationType";
-import { UpdateWellboreLogAction, UpdateWellboreLogsAction } from "../../contexts/navigationStateReducer";
 import { DisplayModalAction, HideContextMenuAction, HideModalAction } from "../../contexts/operationStateReducer";
 import OperationType from "../../contexts/operationType";
 import { ComponentType } from "../../models/componentType";
