+++ Src/WitsmlExplorer.Frontend/components/ContextMenus/WellboreContextMenu.tsx
@@ -2,8 +2,8 @@ import { Typography } from "@equinor/eds-core-react";
 import { Divider, ListItemIcon, MenuItem } from "@material-ui/core";
 import React from "react";
 import { v4 as uuid } from "uuid";
+import { UpdateWellboreAction } from "../../contexts/modificationActions";
 import ModificationType from "../../contexts/modificationType";
-import { UpdateWellboreAction } from "../../contexts/navigationStateReducer";
 import { DisplayModalAction, HideContextMenuAction, HideModalAction } from "../../contexts/operationStateReducer";
 import OperationType from "../../contexts/operationType";
 import { DeleteWellboreJob } from "../../models/jobs/deleteJobs";
