+++ Src/WitsmlExplorer.Frontend/components/Modals/BhaRunPropertiesModal.tsx
@@ -2,8 +2,8 @@ import { Autocomplete } from "@equinor/eds-core-react";
 import { InputAdornment, TextField } from "@material-ui/core";
 import moment from "moment";
 import React, { useEffect, useState } from "react";
+import { UpdateWellboreBhaRunsAction } from "../../contexts/modificationActions";
 import ModificationType from "../../contexts/modificationType";
-import { UpdateWellboreBhaRunsAction } from "../../contexts/navigationStateReducer";
 import { HideModalAction } from "../../contexts/operationStateReducer";
 import OperationType from "../../contexts/operationType";
 import BhaRun from "../../models/bhaRun";
