+++ Src/WitsmlExplorer.Frontend/components/Modals/ServerModal.tsx
@@ -3,8 +3,8 @@ import { Button, TextField } from "@material-ui/core";
 import MuiThumbUpOutlinedIcon from "@material-ui/icons/ThumbUpOutlined";
 import React, { ChangeEvent, useState } from "react";
 import styled from "styled-components";
+import { AddServerAction, RemoveWitsmlServerAction, UpdateServerAction } from "../../contexts/modificationActions";
 import ModificationType from "../../contexts/modificationType";
-import { AddServerAction, RemoveWitsmlServerAction, UpdateServerAction } from "../../contexts/navigationStateReducer";
 import { DisplayModalAction, HideModalAction } from "../../contexts/operationStateReducer";
 import OperationType from "../../contexts/operationType";
 import { Server } from "../../models/server";
