+++ Src/WitsmlExplorer.Frontend/components/Sidebar/ServerManager.tsx
@@ -3,9 +3,10 @@ import { Typography } from "@equinor/eds-core-react";
 import { Divider, FormControl as MuiFormControl, FormHelperText, InputLabel, Link, ListItemIcon, ListItemSecondaryAction, MenuItem, Select } from "@material-ui/core";
 import React, { useContext, useEffect, useState } from "react";
 import styled from "styled-components";
+import { UpdateServerListAction } from "../../contexts/modificationActions";
 import ModificationType from "../../contexts/modificationType";
+import { SelectServerAction } from "../../contexts/navigationActions";
 import NavigationContext from "../../contexts/navigationContext";
-import { SelectServerAction, UpdateServerListAction } from "../../contexts/navigationStateReducer";
 import NavigationType from "../../contexts/navigationType";
 import OperationContext from "../../contexts/operationContext";
 import OperationType from "../../contexts/operationType";
