+++ Src/WitsmlExplorer.Frontend/components/Sidebar/TreeItem.tsx
@@ -4,8 +4,8 @@ import { TreeItem } from "@material-ui/lab";
 import { TreeItemProps } from "@material-ui/lab/TreeItem";
 import React, { useContext } from "react";
 import styled from "styled-components";
+import { ToggleTreeNodeAction } from "../../contexts/navigationActions";
 import NavigationContext from "../../contexts/navigationContext";
-import { ToggleTreeNodeAction } from "../../contexts/navigationStateReducer";
 import NavigationType from "../../contexts/navigationType";
 import { colors } from "../../styles/Colors";
 import Icon from "../../styles/Icons";
