+++ Src/WitsmlExplorer.Frontend/contexts/navigationStateReducer.tsx
@@ -1,18 +1,8 @@
 import { Dispatch, useReducer } from "react";
-import { LogCurveInfoRow } from "../components/ContentViews/LogCurveInfoListView";
-import BhaRun from "../models/bhaRun";
-import LogObject from "../models/logObject";
-import MessageObject from "../models/messageObject";
-import ObjectOnWellbore, { getObjectOnWellboreProperties } from "../models/objectOnWellbore";
+import { getObjectOnWellboreProperties } from "../models/objectOnWellbore";
 import { ObjectType } from "../models/objectType";
-import Rig from "../models/rig";
-import RiskObject from "../models/riskObject";
-import { Server } from "../models/server";
-import Trajectory from "../models/trajectory";
-import Tubular from "../models/tubular";
-import WbGeometryObject from "../models/wbGeometry";
-import Well, { getWellProperties } from "../models/well";
-import Wellbore, {
+import { getWellProperties } from "../models/well";
+import {
   calculateBhaRunGroupId,
   calculateLogGroupId,
   calculateLogTypeId,
@@ -25,79 +15,38 @@ import Wellbore, {
   calculateWellboreNodeId,
   getWellboreProperties
 } from "../models/wellbore";
-import CurveThreshold, { DEFAULT_CURVE_THRESHOLD } from "./curveThreshold";
-import Filter, { EMPTY_FILTER, filterWells } from "./filter";
+import { filterWells } from "./filter";
+import { performModificationAction } from "./modificationStateReducer";
 import ModificationType from "./modificationType";
+import {
+  Action,
+  SelectBhaRunGroupAction,
+  SelectLogCurveInfoAction,
+  SelectLogGroupAction,
+  SelectLogObjectAction,
+  SelectLogTypeAction,
+  SelectMessageGroupAction,
+  SelectMessageObjectAction,
+  SelectRigGroupAction,
+  SelectRiskGroupAction,
+  SelectServerAction,
+  SelectTrajectoryAction,
+  SelectTrajectoryGroupAction,
+  SelectTubularAction,
+  SelectTubularGroupAction,
+  SelectWbGeometryAction,
+  SelectWbGeometryGroupAction,
+  SelectWellAction,
+  SelectWellboreAction,
+  SetCurveThresholdAction,
+  SetFilterAction,
+  ToggleTreeNodeAction
+} from "./navigationActions";
+import { allDeselected, EMPTY_NAVIGATION_STATE, NavigationState, selectedJobsFlag } from "./navigationContext";
 import NavigationType from "./navigationType";
-
-export interface NavigationState {
-  selectedServer: Server;
-  selectedWell: Well;
-  selectedWellbore: Wellbore;
-  selectedBhaRunGroup: string;
-  selectedLogGroup: string;
-  selectedLogTypeGroup: string;
-  selectedLog: LogObject;
-  selectedMessageGroup: string;
-  selectedMessage: MessageObject;
-  selectedRig: Rig;
-  selectedRigGroup: string;
-  selectedRisk: RiskObject;
-  selectedRiskGroup: string;
-  selectedLogCurveInfo: LogCurveInfoRow[];
-  selectedTrajectoryGroup: string;
-  selectedTrajectory: Trajectory;
-  selectedTubularGroup: string;
-  selectedTubular: Tubular;
-  selectedWbGeometryGroup: string;
-  selectedWbGeometry: WbGeometryObject;
-  servers: Server[];
-  currentSelected: Selectable;
-  wells: Well[];
-  filteredWells: Well[];
-  selectedFilter: Filter;
-  selectedCurveThreshold: CurveThreshold;
-  expandedTreeNodes: string[];
-  currentProperties: Map<string, string>;
-}
-
-export type Selectable = Server | Well | Wellbore | string | BhaRun | LogObject | LogCurveInfoRow[] | Trajectory | MessageObject | RiskObject | Rig | WbGeometryObject;
-
-export const selectedJobsFlag = "jobs";
 
 export const initNavigationStateReducer = (): [NavigationState, Dispatch<Action>] => {
   return useReducer(reducer, EMPTY_NAVIGATION_STATE);
-};
-
-export const EMPTY_NAVIGATION_STATE: NavigationState = {
-  selectedServer: null,
-  selectedWell: null,
-  selectedWellbore: null,
-  selectedBhaRunGroup: null,
-  selectedLogGroup: null,
-  selectedLogTypeGroup: null,
-  selectedLog: null,
-  selectedMessageGroup: null,
-  selectedMessage: null,
-  selectedRig: null,
-  selectedRigGroup: null,
-  selectedRiskGroup: null,
-  selectedRisk: null,
-  selectedLogCurveInfo: null,
-  selectedTrajectoryGroup: null,
-  selectedTrajectory: null,
-  selectedTubularGroup: null,
-  selectedTubular: null,
-  selectedWbGeometryGroup: null,
-  selectedWbGeometry: null,
-  servers: [],
-  currentSelected: null,
-  wells: [],
-  filteredWells: [],
-  selectedFilter: EMPTY_FILTER,
-  selectedCurveThreshold: DEFAULT_CURVE_THRESHOLD,
-  expandedTreeNodes: [],
-  currentProperties: new Map<string, string>()
 };
 
 export const reducer = (state: NavigationState, action: Action): NavigationState => {
@@ -161,508 +110,6 @@ const performNavigationAction = (state: NavigationState, action: Action) => {
   }
 };
 
-const performModificationAction = (state: NavigationState, action: Action) => {
-  switch (action.type) {
-    case ModificationType.AddWell:
-      return addWell(state, action);
-    case ModificationType.AddServer:
-      return addServer(state, action);
-    case ModificationType.UpdateServer:
-      return updateServer(state, action);
-    case ModificationType.RemoveWell:
-      return removeWell(state, action);
-    case ModificationType.RemoveWellbore:
-      return removeWellbore(state, action);
-    case ModificationType.RemoveServer:
-      return removeServer(state, action);
-    case ModificationType.UpdateWell:
-      return updateWell(state, action);
-    case ModificationType.AddWellbore:
-      return addWellbore(state, action);
-    case ModificationType.UpdateWellbore:
-      return updateWellbore(state, action);
-    case ModificationType.UpdateBhaRuns:
-      return updateWellboreBhaRuns(state, action);
-    case ModificationType.UpdateLogObjects:
-      return updateWellboreLogs(state, action);
-    case ModificationType.UpdateLogObject:
-      return updateWellboreLog(state, action);
-    case ModificationType.UpdateMessageObjects:
-      return updateWellboreMessages(state, action);
-    case ModificationType.UpdateMessageObject:
-      return updateWellboreMessage(state, action);
-    case ModificationType.UpdateRigsOnWellbore:
-      return updateWellboreRigs(state, action);
-    case ModificationType.UpdateRiskObjects:
-      return updateWellboreRisks(state, action);
-    case ModificationType.UpdateTrajectoryOnWellbore:
-      return updateWellboreTrajectory(state, action);
-    case ModificationType.UpdateTrajectoriesOnWellbore:
-      return updateWellboreTrajectories(state, action);
-    case ModificationType.UpdateTubularOnWellbore:
-      return updateWellboreTubular(state, action);
-    case ModificationType.UpdateTubularsOnWellbore:
-      return updateWellboreTubulars(state, action);
-    case ModificationType.UpdateWbGeometryOnWellbore:
-      return updateWellboreWbGeometry(state, action);
-    case ModificationType.UpdateWbGeometryObjects:
-      return updateWellboreWbGeometrys(state, action);
-    case ModificationType.UpdateServerList:
-      return updateServerList(state, action);
-    case ModificationType.UpdateWells:
-      return updateWells(state, action);
-    default:
-      throw new Error();
-  }
-};
-
-const allDeselected: any = {
-  selectedServer: null,
-  selectedWell: null,
-  selectedWellbore: null,
-  selectedBhaRunGroup: null,
-  selectedLogGroup: null,
-  selectedLogTypeGroup: null,
-  selectedLog: null,
-  selectedMessageGroup: null,
-  selectedMessage: null,
-  selectedRiskGroup: null,
-  selectedRisk: null,
-  selectedLogCurveInfo: null,
-  selectedRig: null,
-  selectedRigGroup: null,
-  selectedTrajectoryGroup: null,
-  selectedTrajectory: null,
-  selectedTubularGroup: null,
-  selectedTubular: null,
-  selectedWbGeometryGroup: null,
-  selectedWbGeometry: null,
-  currentSelected: null,
-  currentProperties: new Map()
-};
-
-export const sortList = (list: Well[] | Wellbore[]): void => {
-  list.sort((a: Well | Wellbore, b: Well | Wellbore) => {
-    const nameA = a.name.toUpperCase();
-    const nameB = b.name.toUpperCase();
-    if (nameA < nameB) {
-      return -1;
-    }
-    if (nameA > nameB) {
-      return 1;
-    }
-    return 0;
-  });
-};
-
-const addServer = (state: NavigationState, { payload }: AddServerAction) => {
-  const { server } = payload;
-  return {
-    ...state,
-    servers: state.servers.concat([server])
-  };
-};
-
-const updateServer = (state: NavigationState, { payload }: UpdateServerAction) => {
-  const { server } = payload;
-  const index = state.servers.findIndex((s) => s.id === server.id);
-  state.servers.splice(index, 1, server);
-  return {
-    ...state,
-    servers: [...state.servers],
-    selectedServer: state.selectedServer?.id === server.id ? server : state.selectedServer,
-    currentSelected: state.selectedServer && state.currentSelected === state.selectedServer ? server : state.currentSelected
-  };
-};
-
-const addWell = (state: NavigationState, { payload }: AddWellAction) => {
-  const { well } = payload;
-  const wells = [...state.wells];
-  wells.push(well);
-  sortList(wells);
-
-  return {
-    ...state,
-    wells,
-    filteredWells: filterWells(wells, state.selectedFilter)
-  };
-};
-
-const updateWell = (state: NavigationState, { payload }: UpdateWellAction) => {
-  const { well } = payload;
-  const wells = [...state.wells];
-  const wellIndex = getWellIndex(wells, well.uid);
-  const { wellbores } = wells[wellIndex];
-
-  wells[wellIndex] = { ...well, wellbores };
-  const refreshedWellIsSelected = state.selectedWell?.uid === well.uid;
-
-  return {
-    ...state,
-    wells,
-    filteredWells: filterWells(wells, state.selectedFilter),
-    selectedWell: refreshedWellIsSelected ? wells[wellIndex] : state.selectedWell
-  };
-};
-
-const addWellbore = (state: NavigationState, { payload }: AddWellboreAction) => {
-  const { wellbore } = payload;
-  const wells = [...state.wells];
-  const wellIndex = getWellIndex(wells, wellbore.wellUid);
-
-  wells[wellIndex].wellbores.push(wellbore);
-  sortList(wells[wellIndex].wellbores);
-
-  return {
-    ...state,
-    wells,
-    filteredWells: filterWells(wells, state.selectedFilter)
-  };
-};
-
-const updateWellbore = (state: NavigationState, { payload }: UpdateWellboreAction) => {
-  const { wellbore } = payload;
-  const wells = [...state.wells];
-  const wellIndex = getWellIndex(wells, wellbore.wellUid);
-  const wellboreIndex = getWellboreIndex(wells, wellIndex, wellbore.uid);
-  const refreshedWellboreIsSelected = state.selectedWellbore?.uid === wellbore.uid;
-  wells[wellIndex].wellbores[wellboreIndex] = { ...wellbore };
-
-  return {
-    ...state,
-    wells,
-    filteredWells: filterWells(wells, state.selectedFilter),
-    selectedWellbore: refreshedWellboreIsSelected ? wells[wellIndex].wellbores[wellboreIndex] : state.selectedWellbore
-  };
-};
-
-const removeWell = (state: NavigationState, { payload }: RemoveWellAction) => {
-  const { wellUid } = payload;
-  const wells = [...state.wells];
-  const wellIndex = getWellIndex(wells, wellUid);
-  wells.splice(wellIndex, 1);
-
-  const selectedIsRemoved = state.selectedWell?.uid === wellUid;
-  const updatedSelectState = selectedIsRemoved
-    ? {
-        ...allDeselected,
-        selectedServer: state.selectedServer
-      }
-    : {};
-  return {
-    ...state,
-    wells,
-    filteredWells: filterWells(wells, state.selectedFilter),
-    ...updatedSelectState
-  };
-};
-
-const removeWellbore = (state: NavigationState, { payload }: RemoveWellboreAction) => {
-  const { wellUid, wellboreUid } = payload;
-  const wells = [...state.wells];
-  const wellIndex = getWellIndex(wells, wellUid);
-  const wellboreIndex = getWellboreIndex(wells, wellIndex, wellboreUid);
-  wells[wellIndex].wellbores.splice(wellboreIndex, 1);
-  sortList(wells[wellIndex].wellbores);
-
-  const selectedIsRemoved = state.selectedWellbore?.uid === wellboreUid;
-  const updatedSelectState = selectedIsRemoved
-    ? {
-        ...allDeselected,
-        selectedServer: state.selectedServer,
-        selectedWell: state.selectedWell
-      }
-    : {};
-  return {
-    ...state,
-    wells,
-    filteredWells: filterWells(wells, state.selectedFilter),
-    ...updatedSelectState
-  };
-};
-
-const removeServer = (state: NavigationState, { payload }: RemoveWitsmlServerAction) => {
-  const { servers, selectedServer } = state;
-  const { serverUid } = payload;
-
-  const index = servers.findIndex((server) => server.id === serverUid);
-  if (index !== -1) {
-    servers.splice(index, 1);
-  }
-
-  const selectedIsRemoved = selectedServer?.id === serverUid;
-  const updatedSelectState = selectedIsRemoved ? { ...allDeselected } : {};
-
-  return {
-    ...state,
-    ...updatedSelectState,
-    servers
-  };
-};
-
-const updateWellboreBhaRuns = (state: NavigationState, { payload }: UpdateWellboreBhaRunsAction) => {
-  const { wells } = state;
-  const { bhaRuns, wellUid, wellboreUid } = payload;
-  const freshWells = replacePropertiesInWellbore(wellUid, wells, wellboreUid, { bhaRuns });
-  return {
-    ...state,
-    ...updateSelectedWellAndWellboreIfNeeded(state, freshWells, wellUid, wellboreUid),
-    wells: freshWells
-  };
-};
-
-const updateWellboreMessages = (state: NavigationState, { payload }: UpdateWellboreMessagesAction) => {
-  const { wells } = state;
-  const { messages, wellUid, wellboreUid } = payload;
-  const freshWells = replacePropertiesInWellbore(wellUid, wells, wellboreUid, { messages });
-  return {
-    ...state,
-    ...updateSelectedWellAndWellboreIfNeeded(state, freshWells, wellUid, wellboreUid),
-    wells: freshWells
-  };
-};
-
-const updateWellboreMessage = (state: NavigationState, { payload }: UpdateWellboreMessageAction) => {
-  const { wells } = state;
-  const { message } = payload;
-  const updatedWells = insertLogIntoWellsStructure(wells, message);
-  const selectedMessage = state.selectedMessage?.uid === message.uid ? message : state.selectedMessage;
-  return {
-    ...state,
-    wells: updatedWells,
-    filteredWells: filterWells(updatedWells, state.selectedFilter),
-    selectedMessage
-  };
-};
-
-const updateWellboreRigs = (state: NavigationState, { payload }: UpdateWellboreRigsAction) => {
-  const { wells } = state;
-  const { rigs, wellUid, wellboreUid } = payload;
-  const freshWells = replacePropertiesInWellbore(wellUid, wells, wellboreUid, { rigs });
-  return {
-    ...state,
-    ...updateSelectedWellAndWellboreIfNeeded(state, freshWells, wellUid, wellboreUid),
-    wells: freshWells
-  };
-};
-
-const updateWellboreRisks = (state: NavigationState, { payload }: UpdateWellboreRisksAction) => {
-  const { wells } = state;
-  const { risks, wellUid, wellboreUid } = payload;
-  const freshWells = replacePropertiesInWellbore(wellUid, wells, wellboreUid, { risks });
-  return {
-    ...state,
-    ...updateSelectedWellAndWellboreIfNeeded(state, freshWells, wellUid, wellboreUid),
-    wells: freshWells
-  };
-};
-
-const updateWellboreLogs = (state: NavigationState, { payload }: UpdateWellboreLogsAction) => {
-  const { wells } = state;
-  const { logs, wellUid, wellboreUid } = payload;
-  const freshWells = replacePropertiesInWellbore(wellUid, wells, wellboreUid, { logs });
-  const calculateGroup = (wellbore: Wellbore) => calculateLogTypeId(wellbore, state.selectedLog.indexType);
-  const { currentSelected, newSelectedObject } = getCurrentSelectedObjectIfRemoved(state, calculateGroup, logs, state.selectedLog, wellboreUid, wellUid);
-  return {
-    ...state,
-    ...updateSelectedWellAndWellboreIfNeeded(state, freshWells, wellUid, wellboreUid),
-    wells: freshWells,
-    currentSelected,
-    selectedLog: newSelectedObject
-  };
-};
-
-const updateWellboreLog = (state: NavigationState, { payload }: UpdateWellboreLogAction) => {
-  const { wells } = state;
-  const { log } = payload;
-  const updatedWells = insertLogIntoWellsStructure(wells, log);
-  const selectedLog = state.selectedLog?.uid === log.uid ? log : state.selectedLog;
-
-  return {
-    ...state,
-    wells: updatedWells,
-    filteredWells: filterWells(updatedWells, state.selectedFilter),
-    selectedLog
-  };
-};
-
-const insertLogIntoWellsStructure = (wells: Well[], log: LogObject): Well[] => {
-  const freshWells = [...wells];
-  const wellIndex = getWellIndex(freshWells, log.wellUid);
-  const wellboreIndex = getWellboreIndex(freshWells, wellIndex, log.wellboreUid);
-  const logIndex = getLogIndex(freshWells, wellIndex, wellboreIndex, log.uid);
-  freshWells[wellIndex].wellbores[wellboreIndex].logs[logIndex] = log;
-
-  return freshWells;
-};
-
-const updateWellboreTrajectory = (state: NavigationState, { payload }: UpdateWellboreTrajectoryAction) => {
-  const { wells } = state;
-  const { trajectory, wellUid, wellboreUid } = payload;
-  const freshWells = [...wells];
-  const wellIndex = getWellIndex(freshWells, wellUid);
-  const wellboreIndex = getWellboreIndex(freshWells, wellIndex, wellboreUid);
-  const freshTrajectories = [...wells[wellIndex].wellbores[wellboreIndex].trajectories];
-  const trajectoryIndex = freshTrajectories.findIndex((t) => t.uid === trajectory.uid);
-  let selectedTrajectory = null;
-  freshTrajectories[trajectoryIndex] = trajectory;
-  selectedTrajectory = state.selectedTrajectory?.uid === trajectory.uid ? trajectory : state.selectedTrajectory;
-  wells[wellIndex].wellbores[wellboreIndex].trajectories = freshTrajectories;
-  return {
-    ...state,
-    wells: freshWells,
-    filteredWells: filterWells(freshWells, state.selectedFilter),
-    selectedTrajectory
-  };
-};
-
-const updateWellboreTrajectories = (state: NavigationState, { payload }: UpdateWellboreTrajectoriesAction) => {
-  const { wells } = state;
-  const { trajectories, wellUid, wellboreUid } = payload;
-  const freshWells = replacePropertiesInWellbore(wellUid, wells, wellboreUid, { trajectories });
-  const { currentSelected, newSelectedObject } = getCurrentSelectedObjectIfRemoved(state, calculateTrajectoryGroupId, trajectories, state.selectedTrajectory, wellboreUid, wellUid);
-  return {
-    ...state,
-    ...updateSelectedWellAndWellboreIfNeeded(state, freshWells, wellUid, wellboreUid),
-    selectedTrajectory: newSelectedObject,
-    currentSelected,
-    wells: freshWells
-  };
-};
-
-const updateWellboreTubulars = (state: NavigationState, { payload }: UpdateWellboreTubularsAction) => {
-  const { wells } = state;
-  const { tubulars, wellUid, wellboreUid } = payload;
-  const freshWells = replacePropertiesInWellbore(wellUid, wells, wellboreUid, { tubulars });
-  const { currentSelected, newSelectedObject } = getCurrentSelectedObjectIfRemoved(state, calculateTubularGroupId, tubulars, state.selectedTubular, wellboreUid, wellUid);
-  return {
-    ...state,
-    ...updateSelectedWellAndWellboreIfNeeded(state, freshWells, wellUid, wellboreUid),
-    selectedTubular: newSelectedObject,
-    currentSelected,
-    wells: freshWells
-  };
-};
-
-const updateWellboreTubular = (state: NavigationState, { payload }: UpdateWellboreTubularAction) => {
-  const { wells } = state;
-  const { tubular, exists } = payload;
-  const freshWells = [...wells];
-  const wellIndex = getWellIndex(freshWells, tubular.wellUid);
-  const wellboreIndex = getWellboreIndex(freshWells, wellIndex, tubular.wellboreUid);
-  const freshTubulars = [...wells[wellIndex].wellbores[wellboreIndex].tubulars];
-  const tubularIndex = freshTubulars.findIndex((t) => t.uid === tubular.uid);
-  let selectedTubular = null;
-  if (exists) {
-    freshTubulars[tubularIndex] = tubular;
-    selectedTubular = state.selectedTubular?.uid === tubular.uid ? tubular : state.selectedTubular;
-  } else {
-    freshTubulars.splice(tubularIndex, 1);
-  }
-  wells[wellIndex].wellbores[wellboreIndex].tubulars = freshTubulars;
-
-  return {
-    ...state,
-    wells: freshWells,
-    filteredWells: filterWells(freshWells, state.selectedFilter),
-    selectedTubular
-  };
-};
-
-const updateWellboreWbGeometrys = (state: NavigationState, { payload }: UpdateWellboreWbGeometrysAction) => {
-  const { wells } = state;
-  const { wbGeometrys, wellUid, wellboreUid } = payload;
-  const freshWells = replacePropertiesInWellbore(wellUid, wells, wellboreUid, { wbGeometrys });
-  return {
-    ...state,
-    ...updateSelectedWellAndWellboreIfNeeded(state, freshWells, wellUid, wellboreUid),
-    wells: freshWells
-  };
-};
-
-const updateWellboreWbGeometry = (state: NavigationState, { payload }: UpdateWellboreWbGeometryAction) => {
-  const { wells } = state;
-  const { wbGeometry, wellUid, wellboreUid } = payload;
-  const freshWells = [...wells];
-  const wellIndex = getWellIndex(freshWells, wellUid);
-  const wellboreIndex = getWellboreIndex(freshWells, wellIndex, wellboreUid);
-  const freshWbGeometries = [...wells[wellIndex].wellbores[wellboreIndex].wbGeometrys];
-  const wbGeometryIndex = freshWbGeometries.findIndex((wbg) => wbg.uid === wbGeometry.uid);
-  let selectedWbGeometry = null;
-  freshWbGeometries[wbGeometryIndex] = wbGeometry;
-  selectedWbGeometry = state.selectedWbGeometry?.uid === wbGeometry.uid ? wbGeometry : state.selectedWbGeometry;
-  wells[wellIndex].wellbores[wellboreIndex].wbGeometrys = freshWbGeometries;
-  return {
-    ...state,
-    wells: freshWells,
-    filteredWells: filterWells(freshWells, state.selectedFilter),
-    selectedWbGeometry: selectedWbGeometry
-  };
-};
-
-//update the current selected object if the current selected object was deleted
-const getCurrentSelectedObjectIfRemoved = (
-  state: NavigationState,
-  calculateGroupId: (wellbore: Wellbore) => string,
-  objects: ObjectOnWellbore[],
-  selectedObject: ObjectOnWellbore,
-  updatedWellboreUid: string,
-  updatedWellUid: string
-) => {
-  const fetchedSelectedObject = objects.find((value) => value.uid === selectedObject?.uid);
-  const isCurrentlySelectedObjectRemoved =
-    state.selectedWell.uid == updatedWellUid &&
-    state.selectedWellbore.uid == updatedWellboreUid && // the update happened on the wellbore that is currently being browsed
-    selectedObject && // there exists a selected object of the same type as the object type that was updated
-    !fetchedSelectedObject && // the selected object does not exist among the objects fetched from the server, implying deletion
-    state.currentSelected == selectedObject; // the object that is currently selected was deleted, requiring update of currently selected object
-  //navigate from the currently selected object to its object group if it was deleted
-  const currentSelected = isCurrentlySelectedObjectRemoved ? calculateGroupId(state.selectedWellbore) : state.currentSelected;
-  return {
-    currentSelected,
-    //update the selected object if it was fetched
-    newSelectedObject: isCurrentlySelectedObjectRemoved ? null : fetchedSelectedObject ?? selectedObject
-  };
-};
-
-const getWellIndex = (wells: Well[], wellUid: string) => {
-  return wells.findIndex((well) => well.uid === wellUid);
-};
-
-const getWellboreIndex = (wells: Well[], wellIndex: number, wellboreUid: string) => {
-  return wells[wellIndex].wellbores.findIndex((wellbore) => wellbore.uid === wellboreUid);
-};
-
-const getLogIndex = (wells: Well[], wellIndex: number, wellboreIndex: number, logUid: string) => {
-  return wells[wellIndex].wellbores[wellboreIndex].logs.findIndex((log) => log.uid === logUid);
-};
-
-const updateSelectedWellAndWellboreIfNeeded = (state: NavigationState, freshWells: Well[], wellUid: string, wellboreUid: string) => {
-  const wellIndex = getWellIndex(freshWells, wellUid);
-  const wellboreIndex = getWellboreIndex(freshWells, wellIndex, wellboreUid);
-  const selectedWell = state.selectedWell?.uid == wellUid ? freshWells[wellIndex] : state.selectedWell;
-  const selectedWellbore = state.selectedWellbore?.uid == wellboreUid ? freshWells[wellIndex].wellbores[wellboreIndex] : state.selectedWellbore;
-  return {
-    selectedWell,
-    selectedWellbore
-  };
-};
-
-const replacePropertiesInWellbore = (
-  wellUid: string,
-  wells: Well[],
-  wellboreUid: string,
-  wellboreProperties: Record<string, BhaRun[] | LogObject[] | Trajectory[] | MessageObject[] | RiskObject[] | Rig[] | WbGeometryObject[]>
-): Well[] => {
-  const wellIndex = getWellIndex(wells, wellUid);
-  const wellboreIndex = getWellboreIndex(wells, wellIndex, wellboreUid);
-  const well = wells[wellIndex];
-  const wellbore = { ...well.wellbores[wellboreIndex], ...wellboreProperties };
-  well.wellbores.splice(wellboreIndex, 1, wellbore);
-  wells.splice(wellIndex, 1, well);
-  return [...wells];
-};
-
 const selectToggleTreeNode = (state: NavigationState, { payload }: ToggleTreeNodeAction) => {
   return {
     ...state,
@@ -682,22 +129,6 @@ const selectServer = (state: NavigationState, { payload }: SelectServerAction) =
     wells: alreadySelected ? state.wells : [],
     filteredWells: alreadySelected ? state.filteredWells : [],
     expandedTreeNodes
-  };
-};
-
-const updateServerList = (state: NavigationState, { payload }: UpdateServerListAction) => {
-  return {
-    ...state,
-    servers: payload.servers
-  };
-};
-
-const updateWells = (state: NavigationState, { payload }: UpdateWellsAction) => {
-  const { wells } = payload;
-  return {
-    ...state,
-    wells: wells,
-    filteredWells: filterWells(wells, state.selectedFilter)
   };
 };
 
@@ -1060,300 +491,3 @@ const toggleTreeNode = (expandedTreeNodes: string[], nodeId: string) => {
     return expandedTreeNodes.filter((expandedNode) => !expandedNode.includes(nodeId));
   }
 };
-
-interface Action {
-  type: any;
-  payload: any;
-}
-
-export interface SelectServerAction extends Action {
-  type: NavigationType.SelectServer;
-  payload: { server: Server };
-}
-
-export interface AddServerAction extends Action {
-  type: ModificationType.AddServer;
-  payload: { server: Server };
-}
-
-export interface UpdateServerAction extends Action {
-  type: ModificationType.UpdateServer;
-  payload: { server: Server };
-}
-
-export interface AddWellAction extends Action {
-  type: ModificationType.AddWell;
-  payload: { well: Well };
-}
-
-export interface UpdateWellAction extends Action {
-  type: ModificationType.UpdateWell;
-  payload: { well: Well };
-}
-
-export interface UpdateWellsAction extends Action {
-  type: ModificationType.UpdateWells;
-  payload: { wells: Well[] };
-}
-
-export interface AddWellboreAction extends Action {
-  type: ModificationType.AddWellbore;
-  payload: { wellbore: Wellbore };
-}
-
-export interface UpdateWellboreAction extends Action {
-  type: ModificationType.UpdateWellbore;
-  payload: { wellbore: Wellbore };
-}
-
-export interface RemoveWellAction extends Action {
-  type: ModificationType.RemoveWell;
-  payload: { wellUid: string };
-}
-
-export interface RemoveWellboreAction extends Action {
-  type: ModificationType.RemoveWellbore;
-  payload: { wellUid: string; wellboreUid: string };
-}
-
-export interface RemoveWitsmlServerAction extends Action {
-  type: ModificationType.RemoveServer;
-  payload: { serverUid: string };
-}
-
-export interface ToggleTreeNodeAction extends Action {
-  type: NavigationType.ToggleTreeNode;
-  payload: { nodeId: string };
-}
-
-export interface UpdateWellboreBhaRunsAction extends Action {
-  type: ModificationType.UpdateBhaRuns;
-  payload: { bhaRuns: BhaRun[]; wellUid: string; wellboreUid: string };
-}
-
-export interface UpdateWellboreLogsAction extends Action {
-  type: ModificationType.UpdateLogObjects;
-  payload: { logs: LogObject[]; wellUid: string; wellboreUid: string };
-}
-
-export interface UpdateWellboreLogAction extends Action {
-  type: ModificationType.UpdateLogObject;
-  payload: { log: LogObject };
-}
-
-export interface UpdateWellboreMessagesAction extends Action {
-  type: ModificationType.UpdateMessageObjects;
-  payload: { messages: MessageObject[]; wellUid: string; wellboreUid: string };
-}
-
-export interface UpdateWellboreMessageAction extends Action {
-  type: ModificationType.UpdateMessageObject;
-  payload: { message: MessageObject };
-}
-
-export interface UpdateWellboreRigsAction extends Action {
-  type: ModificationType.UpdateRigsOnWellbore;
-  payload: { rigs: Rig[]; wellUid: string; wellboreUid: string };
-}
-
-export interface UpdateWellboreRisksAction extends Action {
-  type: ModificationType.UpdateRiskObjects;
-  payload: { risks: RiskObject[]; wellUid: string; wellboreUid: string };
-}
-
-export interface UpdateWellboreRiskAction extends Action {
-  type: ModificationType.UpdateRisksOnWellbore;
-  payload: { risks: RiskObject[]; wellUid: string; wellboreUid: string };
-}
-
-export interface UpdateWellboreTrajectoryAction extends Action {
-  type: ModificationType.UpdateTrajectoryOnWellbore;
-  payload: { trajectory: Trajectory; wellUid: string; wellboreUid: string };
-}
-
-export interface UpdateWellboreTrajectoriesAction extends Action {
-  type: ModificationType.UpdateTrajectoriesOnWellbore;
-  payload: { trajectories: Trajectory[]; wellUid: string; wellboreUid: string };
-}
-
-export interface UpdateWellboreTubularsAction extends Action {
-  type: ModificationType.UpdateTubularsOnWellbore;
-  payload: { tubulars: Tubular[]; wellUid: string; wellboreUid: string };
-}
-
-export interface UpdateWellboreTubularAction extends Action {
-  type: ModificationType.UpdateTubularOnWellbore;
-  payload: { tubular: Tubular; exists: boolean };
-}
-
-export interface UpdateWellboreWbGeometrysAction extends Action {
-  type: ModificationType.UpdateWbGeometryObjects;
-  payload: { wbGeometrys: WbGeometryObject[]; wellUid: string; wellboreUid: string };
-}
-
-export interface UpdateWellboreWbGeometryAction extends Action {
-  type: ModificationType.UpdateWbGeometryOnWellbore;
-  payload: { wbGeometry: WbGeometryObject; wellUid: string; wellboreUid: string };
-}
-
-export interface UpdateServerListAction extends Action {
-  type: ModificationType.UpdateServerList;
-  payload: { servers: Server[] };
-}
-
-export interface SelectWellAction extends Action {
-  type: NavigationType.SelectWell;
-  payload: { well: Well; wellbores: Wellbore[] };
-}
-
-export interface SelectWellboreAction extends Action {
-  type: NavigationType.SelectWellbore;
-  payload: {
-    well: Well;
-    wellbore: Wellbore;
-    bhaRuns: BhaRun[];
-    logs: LogObject[];
-    rigs: Rig[];
-    trajectories: Trajectory[];
-    messages: MessageObject[];
-    risks: RiskObject[];
-    tubulars: Tubular[];
-    wbGeometrys: WbGeometryObject[];
-  };
-}
-
-export interface SelectJobsAction extends Action {
-  type: NavigationType.SelectJobs;
-}
-
-export interface SelectBhaRunGroupAction extends Action {
-  type: NavigationType.SelectBhaRunGroup;
-  payload: { well: Well; wellbore: Wellbore; bhaRunGroup: any };
-}
-
-export interface SelectLogGroupAction extends Action {
-  type: NavigationType.SelectLogGroup;
-  payload: { well: Well; wellbore: Wellbore; logGroup: any };
-}
-
-export interface SelectLogTypeAction extends Action {
-  type: NavigationType.SelectLogType;
-  payload: { well: Well; wellbore: Wellbore; logGroup: string; logTypeGroup: any };
-}
-
-export interface SelectLogObjectAction extends Action {
-  type: NavigationType.SelectLogObject;
-  payload: { log: LogObject; well: Well; wellbore: Wellbore };
-}
-
-export interface SelectMessageGroupAction extends Action {
-  type: NavigationType.SelectMessageGroup;
-  payload: { well: Well; wellbore: Wellbore; messageGroup: any };
-}
-
-export interface SelectMessageObjectAction extends Action {
-  type: NavigationType.SelectMessageObject;
-  payload: { message: MessageObject; well: Well; wellbore: Wellbore };
-}
-export interface SelectRiskGroupAction extends Action {
-  type: NavigationType.SelectRiskGroup;
-  payload: { well: Well; wellbore: Wellbore; riskGroup: any };
-}
-
-export interface SelectLogCurveInfoAction extends Action {
-  type: NavigationType.ShowCurveValues;
-  payload: { logCurveInfo: any };
-}
-
-export interface SelectRigGroupAction extends Action {
-  type: NavigationType.SelectRigGroup;
-  payload: { well: Well; wellbore: Wellbore; rigGroup: any };
-}
-
-export interface SelectTrajectoryGroupAction extends Action {
-  type: NavigationType.SelectTrajectoryGroup;
-  payload: { well: Well; wellbore: Wellbore; trajectoryGroup: any };
-}
-
-export interface SelectTrajectoryAction extends Action {
-  type: NavigationType.SelectTrajectory;
-  payload: { well: Well; wellbore: Wellbore; trajectory: Trajectory; trajectoryGroup: any };
-}
-
-export interface SelectTubularGroupAction extends Action {
-  type: NavigationType.SelectTubularGroup;
-  payload: { well: Well; wellbore: Wellbore; tubularGroup: any };
-}
-
-export interface SelectTubularAction extends Action {
-  type: NavigationType.SelectTubular;
-  payload: { well: Well; wellbore: Wellbore; tubular: Tubular; tubularGroup: any };
-}
-
-export interface SelectWbGeometryGroupAction extends Action {
-  type: NavigationType.SelectWbGeometryGroup;
-  payload: { well: Well; wellbore: Wellbore; wbGeometryGroup: any };
-}
-
-export interface SelectWbGeometryAction extends Action {
-  type: NavigationType.SelectWbGeometry;
-  payload: { well: Well; wellbore: Wellbore; wbGeometry: WbGeometryObject; wbGeometryGroup: any };
-}
-
-export interface SetFilterAction extends Action {
-  type: NavigationType.SetFilter;
-  payload: { filter: Filter };
-}
-
-export interface SetCurveThresholdAction extends Action {
-  type: NavigationType.SetCurveThreshold;
-  payload: { curveThreshold: CurveThreshold };
-}
-
-export type NavigationAction =
-  | AddServerAction
-  | AddWellAction
-  | AddWellboreAction
-  | RemoveWellAction
-  | RemoveWellboreAction
-  | RemoveWitsmlServerAction
-  | UpdateServerAction
-  | UpdateServerListAction
-  | UpdateWellAction
-  | UpdateWellsAction
-  | UpdateWellboreAction
-  | UpdateWellboreBhaRunsAction
-  | UpdateWellboreLogAction
-  | UpdateWellboreLogsAction
-  | UpdateWellboreMessagesAction
-  | UpdateWellboreMessageAction
-  | UpdateWellboreRigsAction
-  | UpdateWellboreRisksAction
-  | UpdateWellboreTrajectoryAction
-  | UpdateWellboreTrajectoriesAction
-  | UpdateWellboreTubularAction
-  | UpdateWellboreTubularsAction
-  | UpdateWellboreWbGeometryAction
-  | UpdateWellboreWbGeometrysAction
-  | ToggleTreeNodeAction
-  | SelectJobsAction
-  | SelectBhaRunGroupAction
-  | SelectLogTypeAction
-  | SelectLogGroupAction
-  | SelectLogCurveInfoAction
-  | SelectLogObjectAction
-  | SelectWellAction
-  | SelectWellboreAction
-  | SelectRigGroupAction
-  | SelectMessageGroupAction
-  | SelectMessageObjectAction
-  | SelectRiskGroupAction
-  | SelectServerAction
-  | SelectTrajectoryAction
-  | SelectTrajectoryGroupAction
-  | SelectTubularAction
-  | SelectTubularGroupAction
-  | SelectWbGeometryGroupAction
-  | SelectWbGeometryAction
-  | SetFilterAction
-  | SetCurveThresholdAction;
